+++ lapdog/cloud/quotas.py
@@ -36,6 +36,16 @@ def quotas(request):
                 401
             )
 
+        if not utils.validate_token(token_info):
+            return (
+                {
+                    'error': 'Rejected token',
+                    'message': 'Token was valid but did not meet Lapdog security requirements. Token must have email, profile, openid, and devstorage.read_write scopes.'
+                    ' Broad users must authenticate via a LapdogToken'
+                },
+                403
+            )
+
         # 2) Check service account
         default_session = utils.generate_default_session(scopes=['https://www.googleapis.com/auth/cloud-platform'])
         account_email = utils.ld_acct_in_project(token_data['email'])