+++ lapdog/cloud/resolution.py
@@ -42,6 +42,16 @@ def insert_resolution(request):
                 401
             )
 
+        if not utils.validate_token(token_info):
+            return (
+                {
+                    'error': 'Rejected token',
+                    'message': 'Token was valid but did not meet Lapdog security requirements. Token must have email, profile, openid, and devstorage.read_write scopes.'
+                    ' Broad users must authenticate via a LapdogToken'
+                },
+                403
+            )
+
         if 'namespace' not in data:
             return (
                 {